fix ci lint
